file structure made and post question page and component
